database is now updated after frontend response edits
the front end now updates and displays updated responses, allowing instructors to update student's responses. TODO: A small bug where the save button updates the database, but does not close the modal, or at least display to the instructor that responses have been saved.  This can be fixed by either closing the modal, or displaying a "responses saved" message temporarily. Also, the save button needs to be styled properly.